+++ zhmcclient_mock/_urihandler.py
@@ -1861,7 +1861,8 @@ class AdapterHandler(GenericGetPropertiesHandler,
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = adapter.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         adapter.manager.remove(adapter.oid)
 
 
@@ -1884,7 +1885,8 @@ class AdapterChangeCryptoTypeHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = adapter.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_required_fields(method, uri, body, ['crypto-type'])
 
         # Check the validity of the new crypto_type
@@ -1919,7 +1921,8 @@ class AdapterChangeAdapterTypeHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = adapter.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_required_fields(method, uri, body, ['type'])
 
         new_adapter_type = body['type']
@@ -2055,7 +2058,8 @@ class PartitionHandler(GenericGetPropertiesHandler,
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                valid_statuses=['stopped'])
@@ -2084,7 +2088,8 @@ class PartitionStartHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                valid_statuses=['stopped'])
@@ -2114,7 +2119,8 @@ class PartitionStopHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                valid_statuses=['active', 'paused',
@@ -2148,7 +2154,8 @@ class PartitionScsiDumpHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                valid_statuses=['active', 'paused',
@@ -2182,7 +2189,8 @@ class PartitionStartDumpProgramHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                valid_statuses=['active', 'degraded', 'paused',
@@ -2215,7 +2223,8 @@ class PartitionPswRestartHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                valid_statuses=['active', 'paused',
@@ -2245,7 +2254,8 @@ class PartitionMountIsoImageHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2295,7 +2305,8 @@ class PartitionUnmountIsoImageHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2350,7 +2361,8 @@ class PartitionIncreaseCryptoConfigHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2393,7 +2405,8 @@ class PartitionDecreaseCryptoConfigHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2437,7 +2450,8 @@ class PartitionChangeCryptoConfigHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2477,7 +2491,8 @@ class HbasHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2532,7 +2547,8 @@ class HbaHandler(GenericGetPropertiesHandler,
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         partition = hba.manager.parent
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2563,7 +2579,8 @@ class HbaReassignPortHandler(object):
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         partition = hmc.lookup_by_uri(partition_uri)  # assert it exists
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2593,7 +2610,8 @@ class NicsHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2669,7 +2687,8 @@ class NicHandler(GenericGetPropertiesHandler,
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         partition = nic.manager.parent
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2696,7 +2715,8 @@ class VirtualFunctionsHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2727,7 +2747,8 @@ class VirtualFunctionHandler(GenericGetPropertiesHandler,
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         partition = vf.manager.parent
         cpc = partition.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
         check_valid_cpc_status(method, uri, cpc)
         check_partition_status(method, uri, partition,
                                invalid_statuses=['starting', 'stopping'])
@@ -2794,7 +2815,8 @@ class VirtualSwitchGetVnicsHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = vswitch.manager.parent
-        assert cpc.dpm_enabled
+        if not cpc.dpm_enabled:
+            raise CpcNotInDpmError(method, uri, cpc)
 
         connected_vnic_uris = vswitch.properties['connected-vnic-uris']
         return {'connected-vnic-uris': connected_vnic_uris}
@@ -3412,7 +3434,8 @@ class LparActivateHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = lpar.manager.parent
-        assert not cpc.dpm_enabled
+        if cpc.dpm_enabled:
+            raise CpcInDpmError(method, uri, cpc)
 
         status = lpar.properties.get('status', None)
         force = body.get('force', False) if body else False
@@ -3474,7 +3497,8 @@ class LparDeactivateHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = lpar.manager.parent
-        assert not cpc.dpm_enabled
+        if cpc.dpm_enabled:
+            raise CpcInDpmError(method, uri, cpc)
 
         status = lpar.properties.get('status', None)
         force = body.get('force', False) if body else False
@@ -3530,7 +3554,8 @@ class LparLoadHandler(object):
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
         cpc = lpar.manager.parent
-        assert not cpc.dpm_enabled
+        if cpc.dpm_enabled:
+            raise CpcInDpmError(method, uri, cpc)
 
         status = lpar.properties.get('status', None)
         force = body.get('force', False) if body else False
@@ -3592,8 +3617,10 @@ class ResetActProfilesHandler(object):
     @staticmethod
     def get(method, hmc, uri, uri_parms, logon_required):
         # pylint: disable=unused-argument
-        """Operation: List Reset Activation Profiles (requires classic
-        mode)."""
+        """
+        Operation: List Reset Activation Profiles.
+        In case of DPM mode, an empty list is returned.
+        """
         cpc_oid = uri_parms[0]
         query_str = uri_parms[1]
         try:
@@ -3602,15 +3629,15 @@ class ResetActProfilesHandler(object):
             new_exc = InvalidResourceError(method, uri)
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
-        assert not cpc.dpm_enabled  # TODO: Verify error or empty result?
         result_profiles = []
-        filter_args = parse_query_parms(method, uri, query_str)
-        for profile in cpc.reset_activation_profiles.list(filter_args):
-            result_profile = {}
-            for prop in profile.properties:
-                if prop in ('element-uri', 'name'):
-                    result_profile[prop] = profile.properties[prop]
-            result_profiles.append(result_profile)
+        if not cpc.dpm_enabled:
+            filter_args = parse_query_parms(method, uri, query_str)
+            for profile in cpc.reset_activation_profiles.list(filter_args):
+                result_profile = {}
+                for prop in profile.properties:
+                    if prop in ('element-uri', 'name'):
+                        result_profile[prop] = profile.properties[prop]
+                result_profiles.append(result_profile)
         return {'reset-activation-profiles': result_profiles}
 
 
@@ -3630,8 +3657,10 @@ class ImageActProfilesHandler(object):
     @staticmethod
     def get(method, hmc, uri, uri_parms, logon_required):
         # pylint: disable=unused-argument
-        """Operation: List Image Activation Profiles (requires classic
-        mode)."""
+        """
+        Operation: List Image Activation Profiles.
+        In case of DPM mode, an empty list is returned.
+        """
         cpc_oid = uri_parms[0]
         query_str = uri_parms[1]
         try:
@@ -3640,15 +3669,15 @@ class ImageActProfilesHandler(object):
             new_exc = InvalidResourceError(method, uri)
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
-        assert not cpc.dpm_enabled  # TODO: Verify error or empty result?
         result_profiles = []
-        filter_args = parse_query_parms(method, uri, query_str)
-        for profile in cpc.image_activation_profiles.list(filter_args):
-            result_profile = {}
-            for prop in profile.properties:
-                if prop in ('element-uri', 'name'):
-                    result_profile[prop] = profile.properties[prop]
-            result_profiles.append(result_profile)
+        if not cpc.dpm_enabled:
+            filter_args = parse_query_parms(method, uri, query_str)
+            for profile in cpc.image_activation_profiles.list(filter_args):
+                result_profile = {}
+                for prop in profile.properties:
+                    if prop in ('element-uri', 'name'):
+                        result_profile[prop] = profile.properties[prop]
+                result_profiles.append(result_profile)
         return {'image-activation-profiles': result_profiles}
 
 
@@ -3668,7 +3697,10 @@ class LoadActProfilesHandler(object):
     @staticmethod
     def get(method, hmc, uri, uri_parms, logon_required):
         # pylint: disable=unused-argument
-        """Operation: List Load Activation Profiles (requires classic mode)."""
+        """
+        Operation: List Load Activation Profiles.
+        In case of DPM mode, an empty list is returned.
+        """
         cpc_oid = uri_parms[0]
         query_str = uri_parms[1]
         try:
@@ -3677,15 +3709,15 @@ class LoadActProfilesHandler(object):
             new_exc = InvalidResourceError(method, uri)
             new_exc.__cause__ = None
             raise new_exc  # zhmcclient_mock.InvalidResourceError
-        assert not cpc.dpm_enabled  # TODO: Verify error or empty result?
         result_profiles = []
-        filter_args = parse_query_parms(method, uri, query_str)
-        for profile in cpc.load_activation_profiles.list(filter_args):
-            result_profile = {}
-            for prop in profile.properties:
-                if prop in ('element-uri', 'name'):
-                    result_profile[prop] = profile.properties[prop]
-            result_profiles.append(result_profile)
+        if not cpc.dpm_enabled:
+            filter_args = parse_query_parms(method, uri, query_str)
+            for profile in cpc.load_activation_profiles.list(filter_args):
+                result_profile = {}
+                for prop in profile.properties:
+                    if prop in ('element-uri', 'name'):
+                        result_profile[prop] = profile.properties[prop]
+                result_profiles.append(result_profile)
         return {'load-activation-profiles': result_profiles}