+++ docs/changes.rst
@@ -65,6 +65,9 @@ Released: 2022-02-17
 
 * Mock support: Fixes for storage groups and added support for storage volumes.
 
+* Mock support: Fixed that operations on activation profiles succeed with an
+  empty result set in case the CPC is in DPM mode, instead of failing.
+
 **Cleanup:**
 
 * Removed the ability to build the Windows executable, triggered by the fact